feat: updated app design for today page too
